Split appencies.tex into separate files;  update figures.
Note that 4_AnalysisbyCoverType is marked as a copy to Appendix-CoverTypeResults because the 7 nonSMCM/X cover type analyses were moved into it.
Also for Appendices.tex to Appendix-FragClassResults.tex - basically the same thing; was reducing the original all-appendices file Appendices.tex into several separate files. This section is large and was copied over.

 Changes to be committed:
	modified:   1_Intro.tex
	modified:   2_Methods.tex
	modified:   3_Results.tex
	modified:   4_AnalysisbyCoverType.tex
	modified:   Appendices.tex
	copied:     4_AnalysisbyCoverType.tex -> Appendix-CoverTypeResults.tex
	new file:   Appendix-DistSeralResults.tex
	copied:     Appendices.tex -> Appendix-FragClassResults.tex
	new file:   Appendix-FragDescriptions.tex
	new file:   Appendix-FragLandResults.tex
	modified:   master.tex